remove unnecessary comments, delete old org-cards component, move ordering component to the result
