+++ src/app/footer/footer.component.scss
