+++ src/app/shared/components/filters-list/categories-filter/categories-filter.component.html
@@ -6,7 +6,7 @@
     <div class="category_details">
       <h3 class="category_title">Спорт</h3>
       <div class="category_link">
-        <a href="#"> 254 гуртки <span class="material-icons arrow">
+        <a routerLink="/result"> 254 гуртки <span class="material-icons arrow">
         arrow_right_alt
         </span>
         </a>
