+++ src/app/shared/components/filters-list/filters-list.component.html
@@ -1,17 +1,5 @@
-<!-- <div fxLayout='row' fxLayoutAlign='space-evenly center' fxLayoutGap='20px'> -->
-<!-- <div fxLayout="row" fxLayoutAlign="space-between center" fxLayoutGap='1rem' class="filters-wrapper"> -->
 <div class="filters-wrapper">
-  <!-- <h1 fxLayoutAlign='start'>Знайди найкращі гуртки у своєму місті!</h1> -->
-  <!-- <app-searchbar></app-searchbar> -->
-  <!-- <div class="shell-search">
-    <app-city-filter class="shell-city"></app-city-filter>
-    <app-searchbar class="shell-searchbar"></app-searchbar>
-  </div> -->
-  <!-- <app-categories-filter></app-categories-filter> -->
   <div>Categories placeholder</div>
-  <!-- <div fxLayout='row' fxLayoutAlign='space-evenly center'> -->
-    <!-- <app-city-filter></app-city-filter> -->
-    <!-- <app-ownership-type-filter></app-ownership-type-filter> -->
   <app-age-filter></app-age-filter>
   <div>Working hours placeholder</div>
   <div>Price placeholder</div>
@@ -19,7 +7,4 @@
     <input type="checkbox">
     Лише з відкритим набором
   </label>
-  <!-- </div>  -->
-
-  <!-- <app-categories-filter></app-categories-filter> -->
 </div>
