--- src/app/shared/components/filters-list/ownership-type-filter/ownership-type-filter.component.html
@@ -1,1 +1,0 @@
-<button mat-stroked-button class="filter__metaData-ownershipToggle">Форма власності</button>
