--- src/app/shared/components/filters-list/ownership-type-filter/ownership-type-filter.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-ownership-type-filter',
-  templateUrl: './ownership-type-filter.component.html',
-  styleUrls: ['./ownership-type-filter.component.scss']
-})
-export class OwnershipTypeFilterComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
