+++ src/app/shared/components/organization-card/organization-card.component.scss
@@ -1,8 +1,5 @@
 .card {
-  // min-width: 274px;
   min-width: 285px;
-  // width: 300px;
-  // height: 440px;
   background: #FFFFFF;
   box-shadow: 0px 6px 16px rgba(0, 0, 0, 0.08);
   border-radius: 5px;
