--- src/app/shared/organization-card/organization-card.component.html
@@ -1,20 +1,0 @@
-<div class="card">
-  <div class="card_header">
-    <div class="card_type">
-      <div class="card_city">{{card.title}}</div>
-      <div class="card_ownership">{{card.title}}</div>
-    </div>
-    <div class="card_rates">
-      <div class="card_price">безкоштовно</div>
-      <div class="card_rate">{{card.title}}</div>
-    </div>
-  </div>
-  <h3 class="card_title">{{card.title}}</h3>
-  <div class="card_footer">
-    <h4 class="card_owner">{{card.title}}</h4>
-    <h5 class="card_age">Вік: </h5>
-    <ul class="card_types">
-      <li class="card_item">{{card.title}}</li>
-    </ul>
-  </div>
-</div>
