--- src/app/shared/organization-card/organization-card.component.scss
@@ -1,72 +1,0 @@
-.card {
-  border: 1px solid black;
-  width: 200px;
-  margin: 15px 3% 0 3%;
-  &_header {
-    height: 90px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    padding: 5px;
-    background-color: #FBAB7E;
-    background-image: linear-gradient(62deg, #FBAB7E 0%, #F7CE68 100%);
-  }
-  &_city, &_ownership, &_price {
-    background-color: grey;
-    color: white;
-    border-radius: 16px;
-    text-align: center;
-    cursor: pointer;
-    transition: 0.3s ease all;
-    &:hover {
-      background-color: rgb(143, 143, 143);
-    }
-  }
-  &_city {
-    width: 65px;
-  }
-  &_ownership {
-    background-color: rgb(66, 66, 255);
-    width: 90px;
-    &:hover {
-      background-color: rgb(0, 0, 255);
-    }
-  }
-  &_price {
-    width: 100px;
-  }
-  &_type {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    gap: 5px;
-  }
-  &_rates {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  &_title {
-    border-top: 1px solid black;
-    border-bottom: 1px solid black;
-    margin: 0;
-    padding: 5px;
-  }
-  &_types {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-  }
-  &_owner {
-    margin: 0;
-  }
-  &_footer {
-    padding: 5px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: flex-start;
-    padding-bottom: 15px;
-    min-height: 120px;
-  }
-}
