--- src/app/shared/organization-card/organization-card.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { OrganizationCardComponent } from './organization-card.component';
-
-describe('OrganizationCardComponent', () => {
-  let component: OrganizationCardComponent;
-  let fixture: ComponentFixture<OrganizationCardComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ OrganizationCardComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(OrganizationCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
