+++ src/app/shared/services/org-cards/org-cards.service.ts
@@ -1,4 +1,4 @@
-import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
@@ -16,16 +16,10 @@ export class OrgCardsService {
 
   constructor(private http: HttpClient) { }
 
-  // setCity(city: string, params: HttpParams){
-  //   if(city){
-  //     return params.set('address.city', city);
-  //   }
-  // }
-  setParams(filters: FilterStateModel): HttpParams {
+  private setParams(filters: FilterStateModel): HttpParams {
     let params = new HttpParams();
     if(filters.searchQuery){
       params = params.set('title', filters.searchQuery);
-      // params = params.set('provider.title', filters.searchQuery);
     }
     if(filters.city){
       params = params.set('address.city', filters.city);
@@ -51,44 +45,7 @@ export class OrgCardsService {
     return this.http.get<orgCard[]>(this.dataUrl, options);
   }
 
-  // getCards(filters: FilterStateModel): Observable<orgCard[]> {
-  //   console.log(filters.toString());
-  //   let params = new HttpParams();
-  //   for(const [key, value] of Object.entries(filters)) {
-  //     if(Array.isArray(value)){
-  //       for(let i = 0; i < value.length; i++){
-  //         params = params.append(key, value[i]);
-  //       }
-  //     } else {
-  //       params = params.set(key, value);
-  //     }
-  //   }
-  //   // if(filters !== null){
-  //   //   if(filters.categories !== null && filters.categories.length > 0){
-  //   //     for(let i = 0; i < filters.categories.length; i++){
-  //   //       params = params.append('category', filters.categories[i]);
-  //   //     }
-  //   //   }
-  //   //   if(filters.city){
-  //   //     params = params.set('city', filters.city);
-  //   //     // console.log("params:" + params.toString());
-  //   //   }
-  //   // }
-  //   // params = params.set('city', 'Kyiv');
-  //   // console.log("params:" + params.toString());
-  //   // params1 = params1.append('city', 'Kyiv');
-  //   // console.log("params1:" + params1.toString());
-  //   // let params = new HttpParams();
- 
-  //   // params = params.append('category', "music");
-  //    const options = { params: params };
-  // //  const options = 
-  //  //   { params: new HttpParams().set('city', filters.city)};
-  // //  console.log(options);
-  //   return this.http.get<orgCard[]>(this.dataUrl, options);
-  // }
-
-  getCards1(): Observable<orgCard[]> {
+  getPopWorkshops(): Observable<orgCard[]> {
     return this.http.get<orgCard[]>(this.dataUrl);
   }
 }
