+++ src/app/shared/shared.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { FiltersListComponent } from './components/filters-list/filters-list.component';
 import { AgeFilterComponent } from './components/filters-list/age-filter/age-filter.component';
 import { SearchbarComponent } from './components/filters-list/searchbar/searchbar.component';
-import { OwnershipTypeFilterComponent } from './components/filters-list/ownership-type-filter/ownership-type-filter.component';
 import { CategoriesFilterComponent } from './components/filters-list/categories-filter/categories-filter.component';
 import { DigitOnlyDirective } from './directives/digit-only.directive';
 import { CityFilterComponent } from './components/filters-list/city-filter/city-filter.component';
@@ -11,7 +10,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MaterialModule } from './material/material.module';
 import { OrganizationCardComponent } from './components/organization-card/organization-card.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
-
+import { RouterModule } from '@angular/router';
 
 
 @NgModule({
@@ -19,7 +18,6 @@ import { FlexLayoutModule } from '@angular/flex-layout';
     FiltersListComponent,
     AgeFilterComponent,
     SearchbarComponent,
-    OwnershipTypeFilterComponent,
     CategoriesFilterComponent,
     CityFilterComponent,
     DigitOnlyDirective,
@@ -29,13 +27,13 @@ import { FlexLayoutModule } from '@angular/flex-layout';
     MaterialModule,
     CommonModule,
     ReactiveFormsModule,
-    FlexLayoutModule
+    FlexLayoutModule,
+    RouterModule
   ],
   exports: [
     FiltersListComponent,
     AgeFilterComponent,
     SearchbarComponent,
-    OwnershipTypeFilterComponent,
     CategoriesFilterComponent,
     CityFilterComponent,
     DigitOnlyDirective,
