+++ src/app/shared/store/filter.actions.ts
@@ -6,12 +6,6 @@ export class setMaxAge {
   static readonly type = '[filter] set max age';
   constructor(public payload: number) {}
 }
-export class getCards {
-  static readonly type = '[filter] get organization cards';
-}
-export class getPopCards {
-  static readonly type = '[filter] get 4 most popular organization cards';
-}
 export class SelectCity {
   static readonly type = '[app] selects city';
   constructor(public payload: string) {}
@@ -20,11 +14,9 @@ export class SetOrder {
   static readonly type = '[filter] Set Order';
   constructor(public payload: string) {}
 }
-
-export class GetActivities {
-  static readonly type = '[filter] Get Activities';
-}
-
 export class GetWorkshops {
   static readonly type = '[filter] Get Workshops';
+}
+export class GetPopWorkshops {
+  static readonly type = '[filter] get 4 most popular organization cards';
 }
