+++ src/app/shared/store/filter.state.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import { State, Action, StateContext, Selector  } from '@ngxs/store';
-import { setMinAge, setMaxAge, SetOrder, SelectCity, getCards, getPopCards, GetActivities, GetWorkshops  } from './filter.actions';
+import { State, Action, StateContext, Selector } from '@ngxs/store';
+import { setMinAge, setMaxAge, SetOrder, SelectCity, GetWorkshops, GetPopWorkshops } from './filter.actions';
 import { OrgCardsService } from 'src/app/shared/services/org-cards/org-cards.service';
 import { orgCard } from '../models/org-card.model';
 import { actCard } from '../models/activities-card.model';
@@ -9,8 +9,6 @@ import { ProviderActivitiesService } from '../services/provider-activities/provi
 export interface FilterStateModel {
   searchQuery: string;
   city: string;
-  stateOwnership: boolean;
-  privateOwnership: boolean;
   isRecruiting: boolean;
   ageFrom: number;
   ageTo: number;
@@ -22,13 +20,11 @@ export interface FilterStateModel {
   name: 'filter',
   defaults: {
     searchQuery: '',
-    city: 'Kyiv',
-    stateOwnership: true,
-    privateOwnership: true,
+    city: '',
     isRecruiting: true,
     ageFrom: 0,
     ageTo: 16,
-    categories: ['art', 'music', 'sport'],
+    categories: [],
     order: 'ratingDesc',
     organizationCards: []
   }
@@ -40,33 +36,32 @@ export class FilterState {
   static orgCards(state: FilterStateModel) {
     return state.organizationCards
   }
+
   constructor(private cardsService: OrgCardsService, private cardsActivitiesService: ProviderActivitiesService){}
 
   @Action(setMinAge)
-    setMinAge({ patchState }: StateContext<FilterStateModel>, { payload }: setMinAge): void {
-      patchState({ ageFrom: payload })
-    }
+  setMinAge({ patchState }: StateContext<FilterStateModel>, { payload }: setMinAge): void {
+    patchState({ ageFrom: payload })
+  }
+
   @Action(setMaxAge)
-    setMaxAge({ patchState }: StateContext<FilterStateModel>, { payload }: setMaxAge): void {
-      patchState({ ageTo: payload })
-    }
+  setMaxAge({ patchState }: StateContext<FilterStateModel>, { payload }: setMaxAge): void {
+    patchState({ ageTo: payload })
+  }
+
   @Action(SelectCity)
   selectCity({ patchState }: StateContext<FilterStateModel>, { payload }: SelectCity): void {
     patchState({ city: payload});
   }
+  
   @Action(SetOrder)
   setOrder({ patchState }: StateContext<FilterStateModel>, { payload }: SetOrder) {
     patchState({ order: payload});
   }
-  @Action(getCards)
-  getCards({ patchState }: StateContext<FilterStateModel>) {
-    return this.cardsService.getCards1().subscribe(
-      (organizationCards: orgCard[]) => patchState({organizationCards})
-    )
-  }
-  @Action(getPopCards)
+
+  @Action(GetPopWorkshops)
   getPopCards({ patchState }: StateContext<FilterStateModel>) {
-    return this.cardsService.getCards1()
+    return this.cardsService.getPopWorkshops()
     .subscribe((organizationCards: orgCard[]) => patchState({organizationCards}))
   }
 
@@ -75,10 +70,4 @@ export class FilterState {
     return this.cardsService.getWorkshops(ctx.getState())
     .subscribe((organizationCards: orgCard[]) => ctx.patchState({organizationCards}))
   }
-
-  // @Action(GetActivities)
-  // getActivities(ctx: StateContext<FilterStateModel>) {
-  //   return this.cardsService.getCards(ctx.getState())
-  //   .subscribe((organizationCards: orgCard[]) => ctx.patchState({organizationCards}))
-  // }
 }
