+++ src/app/shell/main/main.component.html
@@ -38,7 +38,3 @@
     </div>
   </div>
 </div>
-
-
-
-
