+++ src/app/shell/main/result/ordering/ordering.component.scss
