+++ src/app/shell/main/result/organization-cards-list/organization-cards-list.component.html
@@ -1,13 +1,4 @@
-<!-- <div fxLayoutAlign="left center"> -->
-  <!-- <h3>Знайдено Х пропозицій</h3> -->
-  <!-- <app-ordering 
-    [options]="options" 
-    [selectedOption]="selectedOption" 
-    (onChange)="onChangeOrder($event)">
-  </app-ordering> -->
-<!-- </div>   -->
 <div class="cards" fxLayout="row wrap" fxLayoutGap='1.5rem grid' fxLayoutAlign="flex-start" >
-<!-- <div class="cards" fxLayout="row wrap" fxLayoutGap='2%' fxLayoutAlign="flex-start" > -->
   <ng-container> 
     <app-organization-card fxFlex *ngFor="let card of cards" [card]="card"></app-organization-card> 
     <div fxFlex class="placeholder-wrapper"><div class="placeholder"></div></div>
@@ -17,8 +8,4 @@
     <div fxFlex class="placeholder-wrapper"><div class="placeholder"></div></div>
     <div fxFlex class="placeholder-wrapper"><div class="placeholder"></div></div>
   </ng-container>
-  <!-- <div fxFlex class="placeholder"></div> -->
-  <!-- <div fxFlex class="placeholder"></div> -->
-  <!-- <div fxFlex class="placeholder"></div>
-  <div fxFlex class="placeholder"></div> -->
 </div>
