+++ src/app/shell/main/result/organization-cards-list/organization-cards-list.component.scss
@@ -1,5 +1,4 @@
 .placeholder{
   background-color: blue;
   min-width: 285px;
-  // box-sizing: border-box;
 }
