+++ src/app/shell/main/result/organization-cards-list/organization-cards-list.component.ts
@@ -1,15 +1,10 @@
-import {Component, Input, OnInit} from '@angular/core';
-import {Observable} from 'rxjs';
-import {Select, Store} from '@ngxs/store';
-import {FilterState} from '../../../../shared/store/filter.state';
-import {orgCard} from '../../../../shared/models/org-card.model';
-import {GetActivities, SetOrder, getCards, GetWorkshops} from '../../../../shared/store/filter.actions';
+import { Component, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
+import { Select, Store } from '@ngxs/store';
+import { FilterState } from '../../../../shared/store/filter.state';
+import { orgCard } from '../../../../shared/models/org-card.model';
+import { GetWorkshops } from '../../../../shared/store/filter.actions';
 
-
-// export interface Option {
-//   value: string;
-//   viewValue: string;
-// }
 
 @Component({
   selector: 'app-organization-cards-list',
@@ -19,18 +14,8 @@ import {GetActivities, SetOrder, getCards, GetWorkshops} from '../../../../share
 export class OrganizationCardsListComponent implements OnInit {
 
   @Select(FilterState.orgCards) orgCards$: Observable<orgCard[]>;
-  // @Select(state => state.filter.order) order$: Observable<any>;
 
   public cards: orgCard[];
-  // @Input() cards: orgCard[];
-
-  // options: Option[] = [
-  //   {value: 'ratingDesc', viewValue: 'Рейтинг'},
-  //   {value: 'ratingAsc', viewValue: 'Рейтинг'},
-  //   {value: 'priceDesc', viewValue: 'Ціна'},
-  //   {value: 'priceAsc', viewValue: 'Ціна'}
-  // ];
-  // selectedOption: Option;
 
   constructor(private store: Store) {
   }
@@ -38,18 +23,5 @@ export class OrganizationCardsListComponent implements OnInit {
   ngOnInit(): void {
     this.store.dispatch(new GetWorkshops())
     this.orgCards$.subscribe((orgCards: orgCard[]) => this.cards = orgCards)
-
-    // this.order$.subscribe(order => {
-    //   this.selectedOption = this.options.find(option => option.value === order);
-    // });
   }
-
-  // setOrder(order: string) {
-  //   this.store.dispatch(new SetOrder(order));
-  // }
-
-  // onChangeOrder(order: string) {
-  //   this.setOrder(order);
-  // }
-
 }
